[Update] Added Response Entity in transaction controller, description variable in transaction model and update delete method in transaction service

+++ src/main/java/com/financedash/core/service/TransactionService.java
@@ -27,6 +27,11 @@ public class TransactionService {
     }
 
     public void deleteTransactionById(String id){
-        transactionRepository.deleteById(id);
+        if(transactionRepository.existsById(id)){
+            transactionRepository.deleteById(id);
+        }else{
+            throw new TransactionNotFoundException();
+        }
+
     }
 }
